requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.8.0
bleach==6.2.0
bs4==0.0.1
certifi==2019.6.16
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
CnkiSpider==1.1.0
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
et_xmlfile==2.0.0
fake-useragent==0.1.11
fastjsonschema==2.21.1
frozenlist==1.6.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==2.8
ipython==8.12.3
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyterlab_pygments==0.3.0
latexcodec==3.0.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistune==3.1.3
multidict==6.4.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.2.5
openai==1.76.0
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.3
pandocfilters==1.5.1
pillow==11.2.1
pipreqs==0.5.0
playwright==1.51.0
propcache==0.3.1
pybtex==0.24.0
pycparser==2.22
pycurl==7.45.6
pydantic==2.11.3
pydantic_core==2.33.1
pyee==12.1.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PySocks==1.7.1
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2025.2
pywin32==307
PyYAML==6.0.2
referencing==0.36.2
requests==2.22.0
rich==14.0.0
rpds-py==0.25.1
selenium==3.141.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==1.9.2
tencentcloud-sdk-python==3.0.1388
tencentcloud-sdk-python-common==3.0.1388
tinycss2==1.4.0
tqdm==4.33.0
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==1.25.3
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarg==0.1.9
yarl==1.20.0
